move the data sampling and conversion to utils from the potential submodule
